Functional drag/drop schedule
Added drag and drop blocks to the schedule. Blocks can function as individual time blocks or groups of blocks, allowing you to move hours worth of a shift between employees (or simply adjust a shift forward or back a bit).
New blocks still need to be added by the manual controls, but there is already room for a faux 0th employee, so it should be relatively easy.