Faux-Role Blocks
Creation of the Table Key with the faux schedule blocks. Now that I've found a way to read and modify the URL without reloading, it's time to realize that serializing trait objects is a problem I do not have the skill (or at least patience) to solve. Getting traits to work was good practice I think, but it's time to revert to an enum.