Settings done
Settings is now a fully functional context, separate from the business. There are several places where this results in duplication of vital information, such as open and close times, but there is no quick way to ensure the Business will always have access to the Settings, so it will remain that way for now.
Settings are persisted in the URL hash for user convenience.
There is now configuration options for printing. The settings tab also provides a print preview, allowing easy, rapid changes to values like the font size. Printing now also properly removes all other content, so there are no more blank extra pages.
Initialization is now handled properly, with a run-once closure pulling the URL hash to init the business and settings. Turns out, with the previous method, all of the prep work with pulling the hash, parsing it, and generating a new full business was occurring every time Main re-rendered. Glorious.
With the introduction of settings, several more attributes of the Business can be skipped in serde, as they will now be provided by settings (user determined or default).